# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prosthuti API",
    "version": "1.0.0",
    "description": "Local Laravel API. Cookie sessions; mutations require CSRF. See API.md for payloads and workflow constraints. Response schemas are intentionally not exhaustive."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "prosthuti_session"
      }
    }
  },
  "paths": {
...
    "/dashboard": {
      "get": {
        "operationId": "PreparationController_dashboard_get",
        "tags": [
          "dashboard"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": []
      }
    },
...
    "/lessons/{lesson}/bookmark": {
      "put": {
        "operationId": "PreparationController_lessonBookmark_put",
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "lesson",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson}/note": {
      "put": {
        "operationId": "PreparationController_note_put",
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "lesson",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/lessons/{lesson}/progress": {
      "put": {
        "operationId": "PreparationController_progress_put",
        "tags": [
          "lessons"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "lesson",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/questions/{question}/bookmark": {
      "put": {
        "operationId": "PreparationController_questionBookmark_put",
        "tags": [
          "questions"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "question",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/revision": {
      "get": {
        "operationId": "PreparationController_revision_get",
        "tags": [
          "revision"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": []
      }
    },
    "/routine": {
      "get": {
        "operationId": "PreparationController_routine_get",
        "tags": [
          "routine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": []
      },
      "post": {
        "operationId": "PreparationController_saveTask_post",
        "tags": [
          "routine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/routine/{task}": {
      "put": {
        "operationId": "PreparationController_saveTask_put_item",
        "tags": [
          "routine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "PreparationController_deleteTask_delete_item",
        "tags": [
          "routine"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "task",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}